Use actual Scala version in import $ivy-s, add dotty compatibility mode
When running in 2.13 / 3 compatibility mode,
`scala.util.Properties.versionNumberString` returns the 2.13 version,
while we want the 3.x one to be used (unless we're loading a 2.13
dependency in dotty compatibility mode).

We allow 2.13 dependencies to be loaded when running Scala 3, when
"compat" is added to the dependency, like
```
import $ivy.`com.lihaoyi::scalatags:0.7.0 compat`
```

"compat" is added to most import $ivy-s in the tests, until these
dependencies are available for Scala 3.